requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
scikit-learn==1.3.0